+++ api/api-domain/src/main/java/global/fujitsu/api/domain/service/BaseFeeService.java
@@ -1,0 +1,8 @@
+package global.fujitsu.api.domain.service;
+
+import global.fujitsu.api.model.dto.request.BaseFeeRequest;
+import global.fujitsu.api.model.dto.response.BaseFeeResponse;
+
+public interface BaseFeeService {
+    BaseFeeResponse getBaseFee(BaseFeeRequest request);
+}