+++ api/api-domain/src/main/java/global/fujitsu/api/domain/service/WindSpeedFeeService.java
@@ -1,0 +1,8 @@
+package global.fujitsu.api.domain.service;
+
+import global.fujitsu.api.model.dto.request.create.CreateWindSpeedFeeRequest;
+import global.fujitsu.api.model.dto.response.get.WindSpeedFeeResponse;
+import global.fujitsu.api.domain.service.base.BaseService;
+
+public interface WindSpeedFeeService extends BaseService<WindSpeedFeeResponse, CreateWindSpeedFeeRequest> {
+}