+++ api/api-persistence/src/main/java/global/fujitsu/api/entity/model/fee/AirTemperatureFeeEntity.java
@@ -1,0 +1,13 @@
+package global.fujitsu.api.entity.model.fee;
+
+import lombok.NonNull;
+
+import java.math.BigDecimal;
+
+public record AirTemperatureFeeEntity(
+    @NonNull Long id,
+    @NonNull BigDecimal minTemperature,
+    @NonNull BigDecimal maxTemperature,
+    @NonNull BigDecimal fee
+) implements EntityFeeModel {
+}