+++ api/api-persistence/src/main/java/global/fujitsu/api/entity/model/fee/RegionalBaseFeeEntity.java
@@ -1,0 +1,19 @@
+package global.fujitsu.api.entity.model.fee;
+
+import global.fujitsu.api.model.vehicle.VehicleType;
+import global.fujitsu.api.model.region.RegionName;
+import lombok.NonNull;
+
+import java.math.BigDecimal;
+
+public record RegionalBaseFeeEntity(
+    @NonNull Long id,
+    @NonNull RegionName regionName,
+    @NonNull VehicleType vehicleType,
+    @NonNull BigDecimal fee
+) implements EntityFeeModel {
+
+    public RegionalBaseFeeEntity {
+        vehicleType.validate();
+    }
+}