+++ api/api-protocol/src/main/java/global/fujitsu/api/model/dto/request/BaseFeeRequest.java
@@ -1,0 +1,8 @@
+package global.fujitsu.api.model.dto.request;
+
+import global.fujitsu.api.model.region.RegionName;
+import global.fujitsu.api.model.vehicle.VehicleType;
+import lombok.NonNull;
+
+public record BaseFeeRequest(@NonNull RegionName regionName, @NonNull VehicleType vehicleType) {
+}