+++ api/api-protocol/src/main/java/global/fujitsu/api/model/dto/request/create/CreateRequest.java
@@ -1,0 +1,4 @@
+package global.fujitsu.api.model.dto.request.create;
+
+public interface CreateRequest {
+}