+++ api/api-protocol/src/main/java/global/fujitsu/api/model/dto/response/get/AirTemperatureFeeResponse.java
@@ -1,0 +1,13 @@
+package global.fujitsu.api.model.dto.response.get;
+
+import lombok.NonNull;
+
+import java.math.BigDecimal;
+
+public record AirTemperatureFeeResponse(
+    @NonNull Long id,
+    @NonNull BigDecimal minTemperature,
+    @NonNull BigDecimal maxTemperature,
+    @NonNull BigDecimal fee
+) implements GetResponse {
+}