+++ api/api-protocol/src/main/java/global/fujitsu/api/model/dto/response/get/GetResponse.java
@@ -1,0 +1,6 @@
+package global.fujitsu.api.model.dto.response.get;
+
+import global.fujitsu.api.model.dto.response.Response;
+
+public interface GetResponse extends Response {
+}