+++ api/api-protocol/src/main/java/global/fujitsu/api/model/dto/response/get/WeatherPhenomenonFeeResponse.java
@@ -1,0 +1,14 @@
+package global.fujitsu.api.model.dto.response.get;
+
+import global.fujitsu.api.model.vehicle.VehicleType;
+import lombok.NonNull;
+
+import java.math.BigDecimal;
+
+public record WeatherPhenomenonFeeResponse(
+    @NonNull Long id,
+    @NonNull VehicleType vehicleType,
+    @NonNull String weatherPhenomenon,
+    @NonNull BigDecimal price
+) implements GetResponse {
+}