+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@
 ---
 ## [Unreleased] - 21.03.2026
 ### Added
+- Implemented null safe entity models in api-protocol
 - Filled docs(visualization) with Entity Relation Model
 - Designed the Entity Relation model
 - rest-app